third_party/blink/web_tests/fast/js/strict-readonly-statics-expected.txt

This tests that a assigning to a readonly property in a static or symbol table throws in strict mode.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testWindowUndefined() is true
PASS testNumberMAX_VALUE() is true
PASS successfullyParsed is true

TEST COMPLETE
